+++ CHANGELOG.md
@@ -23,6 +23,7 @@
  * Add code generator and related configurations for entity, service, and controller generation
  * Enhance documentation for entity, service, controller, and mapper templates
  * Add comments for entity fields in templates for better clarity
+ * Add support for `@Table` and `@Column` annotations in entity templates
 
 ## 1.0.15